Add new route "logout".
This new route provides a way to invalidate JWTs.

The session in the database are just deleted. I don't think it's very
ethical to track user sessions in this way, so this seems pretty good.

`mensam-server` and `mensam-client` implement the logout.
The frontend is still missing.

There is one case of aynchronous database access in this case, because
the session verification and logout use two separate database
transactions. Imagine two simultaneous logout requests:
  - Logout request 1: Session is verified.
  - Logout request 2: Session is verified.
  - Logout request 1: Session is deleted from database.
  - Logout request 2: Session is already deleted.

In this case there will only be a warning of level `Warn` because this
can totally happen and is not really problematic at all.