Use `StrictData` for database columns.
This fixes some bugs with selda.
It sometimes occured, that when the database was out of sync with the
server's expected schema, then the handler crashed very late.
It was not possible for the 500-throwing catch clause to actually
catch these occurences.

Now many of these cases should be caught again.