Set screen info only once.
This was a wrong assumption.

Previously I thought that the URL push didn't do anything to the actual
model, but it turns out that doing it this way already modifies the
model via `onUrlChange`.

So well, we don't have to do anything else other than relying on
`onUrlChange`.

If we want to get fancy with it, we might want to use `replaceUrl` for
some cases and modify the model ourselves.