Fix space owners.
When the owner column was created we just set some member of the space
as the owner.
This migration will actually set admins to be the owner of the space.

Theretically this is not safe, because there could be admin roles that
are not named 'Admin', but there is no functionality to do so in the
frontend at the moment.