Remove `TimezoneIdentifier`.
This made structuring the code easier.
Now there is no way to create a timezone on the fly.
There is only `timezoneEtcUtc` as a default if it is really necessary.
We are not using `Time.Posix` anymore, just our own `Timezone`.

This led to creating a drop-down menu for selecting a timezone when
creating or editing a space.