Add request tracker to `Model`.
Unfortunately it is not really safe to put the request tracker into the
screen models.
If we would put it into screen models, then we can not really be sure
that we don't cancel requests that are just started after switching the
screen, because they might use the same `Tracker` identifiers.